+++ include/utils/SDL_foward.hpp
@@ -1,0 +1,14 @@
+#ifndef TNT_UTILS_SDL_FORWARD_DECL_HPP
+#define TNT_UTILS_SDL_FORWARD_DECL_HPP
+
+// Utility file containing forward-declaration of some
+// commonly-used opaque SDL2 structures.
+
+typedef struct SDL_Window SDL_Window;
+typedef struct SDL_Renderer SDL_Renderer;
+typedef struct SDL_Texture SDL_Texture;
+
+typedef struct _TTF_Font TTF_Font;
+typedef struct _Mix_Music Mix_Music;
+
+#endif //!TNT_UTILS_SDL_FORWARD_DECL_HPP